+++ Assets/v2/GM.Bounties/GM.Bounties.Models/BountyGameData.cs
@@ -13,7 +13,7 @@ namespace GM.Bounties.Models
         public int UnlockStage;
 
         [JsonRequired]
-        public int HourlyIncome;
+        public long HourlyIncome;
 
         [JsonIgnore]
         public string Name;
